+++ pid/pid_darwin.go
@@ -1,5 +1,10 @@
 package pid
 
+import (
+	"os"
+	"syscall"
+)
+
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
 //                     Copyright (c) 2009-2019 ESSENTIAL KAOS                         //
@@ -7,7 +12,21 @@ package pid
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// IsWorks return if process PID is not -1
+// IsWorks returns true if process with PID from PID file is works
 func IsWorks(name string) bool {
-	return Get(name) != -1
+	pid := Get(name)
+
+	if pid == -1 {
+		return false
+	}
+
+	return IsProcessWorks(pid)
+}
+
+// IsProcessWorks returns true if process with given PID is works
+func IsProcessWorks(pid int) bool {
+	// On Unix systems, FindProcess always succeeds and returns a Process
+	// for the given pid, regardless of whether the process exists.
+	pr, _ := os.FindProcess(pid)
+	return pr.Signal(syscall.Signal(0)) == nil
 }