+++ pid/pid_linux_test.go
@@ -10,7 +10,10 @@ package pid
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 import (
+	"fmt"
 	"io/ioutil"
+	"os"
+	"time"
 
 	. "pkg.re/check.v1"
 )
@@ -23,7 +26,6 @@ func (s *PidSuite) TestIsWorks(c *C) {
 	err := Create("test")
 
 	c.Assert(err, IsNil)
-
 	c.Assert(IsWorks("test"), Equals, true)
 
 	Remove("test")
@@ -31,7 +33,29 @@ func (s *PidSuite) TestIsWorks(c *C) {
 	c.Assert(IsWorks("test"), Equals, false)
 
 	// Write fake pid to pid file
-	ioutil.WriteFile(s.Dir+"/test.pid", []byte("9736163"), 0644)
+	ioutil.WriteFile(s.Dir+"/test.pid", []byte("69999\n"), 0644)
 
 	c.Assert(IsWorks("test"), Equals, false)
+
+	procfsDir = c.MkDir()
+	err = os.Mkdir(procfsDir+"/69999", 0755)
+
+	if err != nil {
+		c.Fatal(fmt.Sprintf("Can't create directory %s", procfsDir+"/69999"))
+	}
+
+	time.Sleep(2 * time.Second)
+
+	err = os.Mkdir(procfsDir+"/1", 0755)
+
+	if err != nil {
+		c.Fatal(fmt.Sprintf("Can't create directory %s", procfsDir+"/1"))
+	}
+
+	c.Assert(IsWorks("test"), Equals, false)
+}
+
+func (s *PidSuite) TestIsProcessWorks(c *C) {
+	c.Assert(IsProcessWorks(os.Getpid()), Equals, true)
+	c.Assert(IsProcessWorks(999999), Equals, false)
 }